backend/requirements.txt
# MarcenAI Backend - Dependências Python

# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Object Storage (MinIO/S3)
minio==7.2.3
boto3==1.34.34

# OpenAI
openai==1.10.0

# Validação e Configuração
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilitários
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pillow==10.2.0

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Testing (opcional, para desenvolvimento)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
